pyproject.toml
[project]
name = "computer-vision-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "scipy>=1.15.2",
    "torch>=2.6.0",
    "transformers>=4.51.3",
]
